move DB module to sibling crate
