+++ Cargo.toml
@@ -1,4 +1,4 @@
 [workspace]
-members = ["proto", "server",]
+members = [ "db", "proto", "server",]
 resolver = "2"
