+++ db/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "db"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+chrono = "0.4.38"
+diesel = { version = "2.2.3", features = ["postgres", "chrono", "serde_json", "uuid"] }
+dotenvy = "0.15.7"
+serde_json = "1.0.127"
+uuid = { version = "1.8.0", features = ["v4"] }
+
