+++ db/src/models/span.rs
@@ -1,9 +1,9 @@
-use crate::db::models::repository::{DieselRepository, Repository};
-use crate::db::schema::span::dsl::span;
+use crate::models::repository::{DieselRepository, Repository};
+use crate::schema::span::dsl::span;
 use diesel::prelude::*;
 
 #[derive(Queryable, Selectable)]
-#[diesel(table_name = crate::db::schema::span)]
+#[diesel(table_name = crate::schema::span)]
 #[diesel(check_for_backend(diesel::pg::Pg))]
 #[allow(dead_code)]
 pub struct Span {
@@ -16,7 +16,7 @@ pub struct Span {
 }
 
 #[derive(Insertable, Selectable, Queryable)]
-#[diesel(table_name = crate::db::schema::span)]
+#[diesel(table_name = crate::schema::span)]
 #[diesel(check_for_backend(diesel::pg::Pg))]
 pub struct InsertableSpan {
     pub ts_start: chrono::DateTime<chrono::Utc>,
@@ -42,7 +42,7 @@ impl<'a> Repository for DieselRepository<'a, span> {
     fn create(&mut self, entity: &Self::InsertableEntity) -> QueryResult<Self::Entity> {
         diesel::insert_into(self.table)
             .values(entity)
-            .returning(crate::db::schema::span::all_columns)
+            .returning(crate::schema::span::all_columns)
             .get_result(self.connection)
     }
