+++ db/src/models/test_registration.rs
@@ -1,9 +1,9 @@
-use crate::db::models::repository::{DieselRepository, Repository};
-use crate::db::schema::test_registration::dsl::test_registration;
+use crate::models::repository::{DieselRepository, Repository};
+use crate::schema::test_registration::dsl::test_registration;
 use diesel::prelude::*;
 
 #[derive(Queryable, Selectable)]
-#[diesel(table_name = crate::db::schema::test_registration)]
+#[diesel(table_name = crate::schema::test_registration)]
 #[diesel(check_for_backend(diesel::pg::Pg))]
 #[allow(dead_code)]
 pub struct TestRegistration {
@@ -15,7 +15,7 @@ pub struct TestRegistration {
 }
 
 #[derive(Insertable, Selectable, Queryable)]
-#[diesel(table_name = crate::db::schema::test_registration)]
+#[diesel(table_name = crate::schema::test_registration)]
 #[diesel(check_for_backend(diesel::pg::Pg))]
 pub struct InsertableTestRegistration {
     pub blob_url: String,
@@ -42,7 +42,7 @@ impl<'a> Repository for DieselRepository<'a, test_registration> {
     fn create(&mut self, entity: &Self::InsertableEntity) -> QueryResult<Self::Entity> {
         diesel::insert_into(self.table)
             .values(entity)
-            .returning(crate::db::schema::test_registration::all_columns)
+            .returning(crate::schema::test_registration::all_columns)
             .get_result(self.connection)
     }
