+++ db/src/models/test_version.rs
@@ -1,9 +1,9 @@
-use crate::db::models::repository::{DieselRepository, Repository};
-use crate::db::schema::test_version::dsl::test_version;
+use crate::models::repository::{DieselRepository, Repository};
+use crate::schema::test_version::dsl::test_version;
 use diesel::prelude::*;
 
 #[derive(Queryable, Selectable)]
-#[diesel(table_name = crate::db::schema::test_version)]
+#[diesel(table_name = crate::schema::test_version)]
 #[diesel(check_for_backend(diesel::pg::Pg))]
 #[allow(dead_code)]
 pub struct TestVersion {
@@ -16,7 +16,7 @@ pub struct TestVersion {
 }
 
 #[derive(Insertable, Selectable, Queryable)]
-#[diesel(table_name = crate::db::schema::test_version)]
+#[diesel(table_name = crate::schema::test_version)]
 #[diesel(check_for_backend(diesel::pg::Pg))]
 pub struct InsertableTestVersion {
     pub name: String,
@@ -44,7 +44,7 @@ impl<'a> Repository for DieselRepository<'a, test_version> {
     fn create(&mut self, entity: &Self::InsertableEntity) -> QueryResult<Self::Entity> {
         diesel::insert_into(self.table)
             .values(entity)
-            .returning(crate::db::schema::test_version::all_columns)
+            .returning(crate::schema::test_version::all_columns)
             .get_result(self.connection)
     }
