+++ server/Cargo.toml
@@ -7,6 +7,7 @@ default-run = "server"
 [dependencies]
 # Local
 olly_proto = { package = "proto", path = "../proto" }
+olly_db = { package = "db", path = "../db" }
 
 # Proto
 tonic = { version= "0.12.0", features = ["prost"] }
