--- server/src/lib.rs
@@ -1,4 +1,0 @@
-pub mod db;
-pub mod queue;
-pub mod register;
-pub mod tracing;
